pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support-triage"
version = "1.0.0"
description = "OpenEnv customer support triage environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.12",
  "uvicorn==0.34.0",
  "pydantic==2.11.3",
  "openai==1.76.0",
  "python-dotenv==1.0.1",
  "openenv>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["openenv_support_triage", "app", "server"]
